Add tendermint consensus
Problem:
We want to be distributed and decentralised but we also want unique
names. This requires distributed consensus.

Solution:
Plumb events through tendermint. Tenderming is a BFT consensus
alogorithm that consolidates transactions into blocks and resolves them
on the order of once a second. This is easily interactive enough for a
twitter-alike.